+++ BlueprintUICommonControls/Sources/TransitionContainer.swift
@@ -6,14 +6,22 @@ import UIKit
 /// disappears, or changes layout.
 public struct TransitionContainer: Element {
 
-    public var appearingTransition: VisibilityTransition = .fade
-    public var disappearingTransition: VisibilityTransition = .fade
-    public var layoutTransition: LayoutTransition = .specific(AnimationAttributes())
+    public var appearingTransition: VisibilityTransition
+    public var disappearingTransition: VisibilityTransition
+    public var layoutTransition: LayoutTransition
 
     public var wrappedElement: Element
 
-    public init(wrapping element: Element) {
+    public init(
+        wrapping element: Element,
+        appearingTransition: VisibilityTransition = .fade,
+        disappearingTransition: VisibilityTransition = .fade,
+        layoutTransition: LayoutTransition = .specific(AnimationAttributes())
+    ) {
         self.wrappedElement = element
+        self.appearingTransition = appearingTransition
+        self.disappearingTransition = disappearingTransition
+        self.layoutTransition = layoutTransition
     }
 
     public var content: ElementContent {
@@ -28,4 +36,39 @@ public struct TransitionContainer: Element {
         }
     }
 
+}
+
+public extension Element {
+
+    /// Wraps the element in a transition container to provide an animated transition.
+    ///
+    /// - Parameters:
+    ///   - appear: The transition to use when the element appears. By default, `.none`.
+    ///   - disappear: The transition to use when the element disappears. By default, `.none`.
+    ///   - layout: The animation to use when the element changes layout. By default, `.none`.
+    func transition(
+        appear: VisibilityTransition = .none,
+        disappear: VisibilityTransition = .none,
+        layout: LayoutTransition = .none
+    ) -> TransitionContainer {
+        TransitionContainer(
+            wrapping: self,
+            appearingTransition: appear,
+            disappearingTransition: disappear,
+            layoutTransition: layout
+        )
+    }
+
+    /// Wraps the element in a transition container to provide an animated transition when its visibility changes.
+    ///
+    /// - Parameters:
+    ///   - appearAndDisappear: The transition to use when the element appears and disappears.
+    func transition(appearAndDisappear: VisibilityTransition) -> TransitionContainer {
+        TransitionContainer(
+            wrapping: self,
+            appearingTransition: appearAndDisappear,
+            disappearingTransition: appearAndDisappear,
+            layoutTransition: .none
+        )
+    }
 }